Handle fragemented CITP packets when they fragment on a data block read
This adds a new helper method to both CITP and TCNet binary readers that throws and EndOfStreamException when not able to read all the requested data. The default ReadBytes method does not throw EndOfStream so silently reads insifficient data from the TCP stream.

It is important that a CITP packet read throws EndOfStreamException when it reaches the end of the stream as this indicates that we must wait for further data to be read from the stream. In most cases if the TCP packet fragments this will be handled by the Windows API routines but of a sender has incorrectly fragmented the packet this will result in the API call returning a short packet. If this ocurrs in the middle of reading byte data then a successfull packet is read but contains insufficient data.

Alsso adds a test to simulate packet fragmentation and test that the packets are built correctly.